Extract tracker to a differnet file
